refactor: Move Luis functionalities to Recognizer namespace
