+++ src/Common/Recognizer/CommonRecognizer.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Luis;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.AI.Luis;
 using Microsoft.Extensions.Configuration;
 using LuisPredictionOptions = Microsoft.Bot.Builder.AI.LuisV3.LuisPredictionOptions;
 
-namespace Bot.Common
+namespace Bot.Common.Recognizer
 {
     public sealed class CommonRecognizer : IRecognizer
     {
