+++ src/DIC/DicHandler.cs
@@ -1,10 +1,10 @@
 ﻿using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Bot.Common.Recognizer;
 using Bot.Data;
 using Bot.States;
 using Bot.Supports;
-using Luis;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Schema;
